eigrpd: Create eigrp_cli.h and move forward declarations
Move the forward declarations for eigrp_cli.c from eigrpd.h
into eigrp_cli.h.
