+++ eigrpd/eigrpd.h
@@ -76,50 +76,6 @@ extern struct eigrp *eigrp_get(uint16_t as, vrf_id_t vrf_id);
 extern struct eigrp *eigrp_lookup(vrf_id_t vrf_id);
 extern void eigrp_router_id_update(struct eigrp *eigrp);
 
-/* eigrp_cli.c */
-extern void eigrp_cli_show_header(struct vty *vty, struct lyd_node *dnode,
-				  bool show_defaults);
-extern void eigrp_cli_show_end_header(struct vty *vty, struct lyd_node *dnode);
-extern void eigrp_cli_show_router_id(struct vty *vty, struct lyd_node *dnode,
-				     bool show_defaults);
-extern void eigrp_cli_show_passive_interface(struct vty *vty,
-					     struct lyd_node *dnode,
-					     bool show_defaults);
-extern void eigrp_cli_show_active_time(struct vty *vty, struct lyd_node *dnode,
-				       bool show_defaults);
-extern void eigrp_cli_show_variance(struct vty *vty, struct lyd_node *dnode,
-				    bool show_defaults);
-extern void eigrp_cli_show_maximum_paths(struct vty *vty,
-					 struct lyd_node *dnode,
-					 bool show_defaults);
-extern void eigrp_cli_show_metrics(struct vty *vty, struct lyd_node *dnode,
-				   bool show_defaults);
-extern void eigrp_cli_show_network(struct vty *vty, struct lyd_node *dnode,
-				   bool show_defaults);
-extern void eigrp_cli_show_neighbor(struct vty *vty, struct lyd_node *dnode,
-				    bool show_defaults);
-extern void eigrp_cli_show_redistribute(struct vty *vty,
-					struct lyd_node *dnode,
-					bool show_defaults);
-extern void eigrp_cli_show_delay(struct vty *vty, struct lyd_node *dnode,
-				 bool show_defaults);
-extern void eigrp_cli_show_bandwidth(struct vty *vty, struct lyd_node *dnode,
-				     bool show_defaults);
-extern void eigrp_cli_show_hello_interval(struct vty *vty,
-					  struct lyd_node *dnode,
-					  bool show_defaults);
-extern void eigrp_cli_show_hold_time(struct vty *vty, struct lyd_node *dnode,
-				     bool show_defaults);
-extern void eigrp_cli_show_summarize_address(struct vty *vty,
-					     struct lyd_node *dnode,
-					     bool show_defaults);
-extern void eigrp_cli_show_authentication(struct vty *vty,
-					  struct lyd_node *dnode,
-					  bool show_defaults);
-extern void eigrp_cli_show_keychain(struct vty *vty, struct lyd_node *dnode,
-				    bool show_defaults);
-extern void eigrp_cli_init(void);
-
 /* eigrp_northbound.c */
 extern const struct frr_yang_module_info frr_eigrpd_info;
